feat(decorators)!: modern stage 3 accessors, legacy moves to envapt/legacy
