+++ packages/envapt/package.json
@@ -52,6 +52,36 @@
                 "types": "./dist/node/config.d.cts",
                 "default": "./dist/node/config.cjs"
             }
+        },
+        "./legacy": {
+            "workerd": {
+                "types": "./dist/workerd/legacy.d.mts",
+                "default": "./dist/workerd/legacy.mjs"
+            },
+            "browser": {
+                "types": "./dist/browser/legacy.d.mts",
+                "default": "./dist/browser/legacy.mjs"
+            },
+            "node": {
+                "import": {
+                    "types": "./dist/node/legacy.d.mts",
+                    "default": "./dist/node/legacy.mjs"
+                },
+                "require": {
+                    "types": "./dist/node/legacy.d.cts",
+                    "default": "./dist/node/legacy.cjs"
+                }
+            },
+            "default": {
+                "import": {
+                    "types": "./dist/node/legacy.d.mts",
+                    "default": "./dist/node/legacy.mjs"
+                },
+                "require": {
+                    "types": "./dist/node/legacy.d.cts",
+                    "default": "./dist/node/legacy.cjs"
+                }
+            }
         }
     },
     "sideEffects": [
@@ -91,6 +121,7 @@
         "test:browser": "tsc --noEmit -p tests/browser/tsconfig.json && vitest run --config vitest.browser.config.ts",
         "test:consumer-build": "node tests/consumer-build/run.mjs",
         "test:tsc-emit": "node tests/tsc-emit/run.mjs",
+        "test:stage3-emit": "node tests/stage3-emit/run.mjs",
         "test:all": "node ../../scripts/test-all.mjs",
         "coverage": "pnpm run test --coverage",
         "cs": "changeset",
