+++ packages/envapt/src/browser.ts
@@ -1,7 +1,8 @@
-import { Envapter } from './Envapter';
-import { installFileApiStubs } from './installFileApiStubs';
+import { Envapter } from './engine/Envapter';
+import { installFileApiStubs } from './engine/installFileApiStubs';
 
 export * from './common';
+export * from './decorators/modern';
 export { Envapter };
 
 installFileApiStubs();
