+++ packages/envapt/src/common.ts
@@ -1,14 +1,11 @@
-// Runtime-agnostic public surface shared by every entry barrel (node/browser/workerd). Sources are
-// imported by file (not via ./sources) so this stays free of NodeEnvSource's node:* graph.
-export { Envapt, EnvBool, EnvNum, EnvStr, EnvTime, EnvUrl } from './decorators';
-export { Environment } from './Envapter';
+export { Environment } from './engine/Envapter';
 export { Converters, isArrayOf } from './converters';
 export { ManualEnvSource } from './sources/ManualEnvSource';
 export { WorkerEnvSource } from './sources/WorkerEnvSource';
 export type { ArrayElement, ArrayOf, ConverterToken, CustomElementConverter } from './converters';
-export type { DebugLevel } from './Debug';
-export type { EnvFileOptions } from './Dotenv';
-export type { InferSchemaInput, InferSchemaOutput, StandardSchemaV1 } from './StandardSchema';
-export * from './Error';
+export type { DebugLevel } from './infra/Debug';
+export type { EnvFileOptions } from './infra/Dotenv';
+export type { InferSchemaInput, InferSchemaOutput, StandardSchemaV1 } from './infra/StandardSchema';
+export * from './infra/Error';
 
 export type * from './types';
