+++ packages/envapt/src/index.ts
@@ -1,3 +1,4 @@
 export * from './common';
-export { NodeEnvapter as Envapter } from './NodeEnvapter';
+export * from './decorators/modern';
+export { NodeEnvapter as Envapter } from './engine/NodeEnvapter';
 export { NodeEnvSource } from './sources/NodeEnvSource';
