+++ packages/envapt/src/infra/Error.ts
