+++ packages/envapt/src/legacy.ts
@@ -1,0 +1,2 @@
+export * from './decorators/legacy';
+export type { EnvaptFieldDecorator } from './types';
